testsuite/mk/boilerplate.mk: make CHECK_API_ANNOTATIONS and CHECK_PPR overridable
Without this patch, boilerplate.mk (which is included by a lot of Makefiles from
our testsuite) just assumes they reside in the usual inplace directory, which is
not friendly to hadrian and this makes a lot of tests (e.g T10255) fail when
building GHC and running the testsuite with hadrian.

With this patch, the said tests pass.