rts/Linker.c: define the 'mmap_again' label only for x86_64
Summary:
We only ever goto mmap_again when the host arch is x86_64, which
made the build fail on i386, complaining that mmap_again is defined but not
used anywhere. This patches simply moves the label definition inside the
`#if defined(x86_64_HOST_ARCH)` block.

Test Plan: Build the RTS on i386

Reviewers: bgamari, erikd, simonmar

Subscribers: rwbarton, carter

Differential Revision: https://phabricator.haskell.org/D5230