hadrian: make it possible to run the testsuite with quickest and quick
More generally, we so far assumed that the testsuite would be executed with
a flavour that's as comprehensive as perf in terms of available RTS and
library flavours (at least vanilla + dynamic + prof). This would manifest
itself concretely by needing 3 "ways" of the iserv program,
unconditionally.

We now only require the ways among vanilla, dynamic and prof that we can find
in our current Flavour's rtsWays.